+++ cmd/assistant.go
@@ -9,11 +9,11 @@ var assistantCmd = &cobra.Command{
 	Aliases: []string{"assist", "a"},
 	Short:   "Group of commands related to `robot assistant`.",
 	Long: `This set of commands relate to Robocorp Robot Assistant related tasks.
-They are either local, or in relation to Robocorp Cloud and Robocorp App.`,
+They are either local, or in relation to Robocorp Control Room and Robocorp App.`,
 }
 
 func init() {
 	rootCmd.AddCommand(assistantCmd)
 
-	assistantCmd.PersistentFlags().StringVarP(&accountName, "account", "", "", "Account used for Robocorp Cloud operations.")
+	assistantCmd.PersistentFlags().StringVarP(&accountName, "account", "", "", "Account used for Robocorp Control Room operations.")
 }