+++ cmd/cloud.go
@@ -7,12 +7,12 @@ import (
 var cloudCmd = &cobra.Command{
 	Use:     "cloud",
 	Aliases: []string{"robocorp", "c"},
-	Short:   "Group of commands related to `Robocorp Cloud`.",
-	Long:    `This group of commands apply to communication with Robocorp Cloud.`,
+	Short:   "Group of commands related to `Robocorp Control Room`.",
+	Long:    `This group of commands apply to communication with Robocorp Control Room.`,
 }
 
 func init() {
 	rootCmd.AddCommand(cloudCmd)
 
-	cloudCmd.PersistentFlags().StringVarP(&accountName, "account", "a", "", "Account used for Robocorp Cloud operations.")
+	cloudCmd.PersistentFlags().StringVarP(&accountName, "account", "a", "", "Account used for Robocorp Control Room operations.")
 }