+++ cmd/metric.go
@@ -17,8 +17,8 @@ var (
 
 var metricCmd = &cobra.Command{
 	Use:   "metric",
-	Short: "Send some metric to Robocorp Cloud.",
-	Long:  "Send some metric to Robocorp Cloud.",
+	Short: "Send some metric to Robocorp Control Room.",
+	Long:  "Send some metric to Robocorp Control Room.",
 	Run: func(cmd *cobra.Command, args []string) {
 		if common.DebugFlag {
 			defer common.Stopwatch("Feedback metric lasted").Report()