+++ cmd/pull.go
@@ -15,8 +15,8 @@ import (
 
 var pullCmd = &cobra.Command{
 	Use:   "pull",
-	Short: "Pull a robot from Robocorp Cloud and unwrap it into local directory.",
-	Long:  "Pull a robot from Robocorp Cloud and unwrap it into local directory.",
+	Short: "Pull a robot from Robocorp Control Room and unwrap it into local directory.",
+	Long:  "Pull a robot from Robocorp Control Room and unwrap it into local directory.",
 	Run: func(cmd *cobra.Command, args []string) {
 		if common.DebugFlag {
 			defer common.Stopwatch("Pull lasted").Report()