+++ cmd/robot.go
@@ -8,8 +8,8 @@ var robotCmd = &cobra.Command{
 	Use:     "robot",
 	Aliases: []string{"r"},
 	Short:   "Group of commands related to `robot`.",
-	Long: `This set of commands relate to Robocorp Cloud related tasks. They are
-executed either locally, or in connection to Robocorp Cloud and Robocorp App.`,
+	Long: `This set of commands relate to Robocorp Control Room related tasks. They are
+executed either locally, or in connection to Robocorp Control Room and Robocorp App.`,
 }
 
 func init() {