+++ cmd/task.go
@@ -8,8 +8,8 @@ var taskCmd = &cobra.Command{
 	Use:     "task",
 	Aliases: []string{"t"},
 	Short:   "Group of commands related to `task`.",
-	Long: `This set of commands relate to Robocorp Cloud related tasks. They are
-executed either locally, or in connection to Robocorp Cloud and Robocorp App.`,
+	Long: `This set of commands relate to Robocorp Control Room related tasks. They are
+executed either locally, or in connection to Robocorp Control Room and Robocorp App.`,
 }
 
 func init() {